Add tests Rename project
