+++ pyproject.toml
@@ -1,17 +1,17 @@
 [tool.poetry]
-name = "py-image-uploader"
+name = "image-upload-cli"
 version = "0.2.1"
 description = "Upload images via APIs"
 authors = ["DeadNews <uhjnnn@gmail.com>"]
 license = "MIT"
 readme = "README.md"
-homepage = "https://github.com/DeadNews/py-image-uploader"
-repository = "https://github.com/DeadNews/py-image-uploader"
+homepage = "https://github.com/DeadNews/image-upload-cli"
+repository = "https://github.com/DeadNews/image-upload-cli"
 keywords = ["cli", "image", "uploader"]
 classifiers = ["Environment :: Console", "Operating System :: OS Independent"]
 
 [tool.poetry.scripts]
-py-image-uploader = "py_image_uploader.cli:cli"
+image-upload-cli = "image_upload_cli.cli:cli"
 
 [tool.poetry.dependencies]
 python = "^3.9"
