+++ tests/test_cli.py
@@ -1,9 +1,14 @@
 #!/usr/bin/env python
 from click.testing import CliRunner
 
-from src.py_image_uploader.cli import cli
+from src.image_upload_cli.cli import cli
 
 
 def test_click():
     runner = CliRunner()
-    assert runner.invoke(cli=cli, args=["--help"]).exit_code == 0
+    assert runner.invoke(cli, ["--help"]).exit_code == 0
+
+
+def test_uploadcare():
+    runner = CliRunner()
+    assert runner.invoke(cli, ["tests/pixel.png", "-h", "uploadcare"]).exit_code == 0
